Migrate contact form.
Change the behavior. Since we are using only front end tech, use a
simple js mechanism to open the user's email client, instead of sending
to the email address on the server. Also simplify the fields. Not the
best approach, but this is what can be done, and should be fine for now.